pyproject.toml
[tool.poetry]
name = "network-api"
version = "0.1.0"
description = "An API endpoint for interacting with your network"
authors = ["Sudarshan V"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.78.0"}
napalm = "^3.4.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.6.0"
flake8 = "^4.0.1"
yamllint = "^1.27.1"
pylama = "^8.3.8"
bandit = "^1.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
